# crates/bunker-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LAN BUNKER API",
    "description": "The LAN BUNKER backend over SQLite: players and auth, tournaments and brackets, events and check-ins, and the cycles ledger",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/admin/events": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "list_all_events",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Every event, drafts included, the latest night first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_Event"
                }
              }
            }
          },
          "400": {
            "description": "A query parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin, or `PasswordChangeRequired`: a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "create_event",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventFields"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "A new draft with its own check-in code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "The body is not valid JSON, or a parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin, or `PasswordChangeRequired`: a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "The body is too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "415": {
            "description": "The body is not `application/json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "A field out of range, or the end before the start",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/events/{id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "get_event",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EventId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The event, its check-in code and who came",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetail"
                }
              }
            }
          },
          "400": {
            "description": "A path parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin, or `PasswordChangeRequired`: a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "admin"
        ],
        "operationId": "update_event",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EventId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventFields"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Every field replaced. The status and the code stay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "The body is not valid JSON, or a parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin, or `PasswordChangeRequired`: a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "The body is too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "415": {
            "description": "The body is not `application/json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "A field out of range, or the end before the start",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "operationId": "delete_event",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EventId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Gone, with its check-ins and their cycles"
          },
          "400": {
            "description": "A path parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin, or `PasswordChangeRequired`: a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/events/{id}/checkins": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "add_checkin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EventId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckinAdd"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Already in. The time of the first check-in, and zero cycles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinReceipt"
                }
              }
            }
          },
          "201": {
            "description": "Checked in by hand, cycles paid. Any status, any time",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinReceipt"
                }
              }
            }
          },
          "400": {
            "description": "The body is not valid JSON, or a parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin, or `PasswordChangeRequired`: a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Unknown event or player",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "The body is too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "415": {
            "description": "The body is not `application/json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "A field is missing, unknown or out of range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/events/{id}/status": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "change_event_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EventId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventStatusChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Published or back to draft. A draft closes its door",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "The body is not valid JSON, or a parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin, or `PasswordChangeRequired`: a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "The body is too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "415": {
            "description": "The body is not `application/json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "A field is missing, unknown or out of range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/me/checkins": {
      "get": {
        "tags": [
          "events"
        ],
        "operationId": "checkins",
        "responses": {
          "200": {
            "description": "The events the caller checked in to",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Checkins"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "`PasswordChangeRequired`: the caller must choose a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "The one shape every failure has on the wire.",
        "required": [
          "code",
          "message",
          "status"
        ],
        "properties": {
          "cause": {
            "type": [
              "string",
              "null"
            ],
            "description": "The chain of causes. It is absent if a user can see the response."
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "CheckinAdd": {
        "type": "object",
        "description": "Body of `POST /api/admin/events/{id}/checkins`: an admin checks a player in\nby hand, for a night that is over or a phone that did not scan.",
        "required": [
          "playerId"
        ],
        "properties": {
          "playerId": {
            "$ref": "#/components/schemas/PlayerId"
          }
        },
        "additionalProperties": false
      },
      "CheckinCode": {
        "type": "string",
        "maxLength": 12,
        "minLength": 12,
        "pattern": "^[a-z0-9]{12}$"
      },
...
      "CheckinReceipt": {
        "type": "object",
        "description": "Answer of `POST /api/checkin/{code}`. A repeat answers the time of the\nfirst check-in and pays nothing.",
        "required": [
          "event",
          "checkedInAt",
          "cycles"
        ],
        "properties": {
          "checkedInAt": {
            "type": "string",
            "format": "date-time"
          },
          "cycles": {
            "type": "integer",
            "format": "int64",
            "description": "What this call paid: the check-in cycles the first time, zero after."
          },
          "event": {
            "$ref": "#/components/schemas/Event"
          }
        }
      },
...
      "Checkins": {
        "type": "object",
        "description": "Answer of `GET /api/me/checkins`: the events the caller checked in to.",
        "required": [
          "events"
        ],
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventId"
            }
          }
        }
      },
...
      "Description": {
        "type": "string",
        "maxLength": 1000,
        "minLength": 0
      },
...
      "Event": {
        "type": "object",
        "description": "An event as every client sees it. The check-in code is not here: only an\nadmin reads it, through [`EventDetail`].",
        "required": [
          "id",
          "name",
          "location",
          "games",
          "description",
          "startsAt",
          "endsAt",
          "status",
          "checkinCount",
          "createdAt"
        ],
        "properties": {
          "checkinCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "$ref": "#/components/schemas/Description"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time",
            "description": "The last game."
          },
          "games": {
            "$ref": "#/components/schemas/Games"
          },
          "id": {
            "$ref": "#/components/schemas/EventId"
          },
          "image": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ImageName"
              }
            ]
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "name": {
            "$ref": "#/components/schemas/EventName"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time",
            "description": "The doors open."
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          }
        }
      },
      "EventDetail": {
        "type": "object",
        "description": "Answer of `GET /api/admin/events/{id}`: the event, the code for the QR at\nthe door, and everyone who came.",
        "required": [
          "event",
          "checkinCode",
          "checkins"
        ],
        "properties": {
          "checkinCode": {
            "$ref": "#/components/schemas/CheckinCode"
          },
          "checkins": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Checkin"
            },
            "description": "First at the door first."
          },
          "event": {
            "$ref": "#/components/schemas/Event"
          }
        }
      },
      "EventFields": {
        "type": "object",
        "description": "Body of `POST /api/admin/events` and of `PUT /api/admin/events/{id}`. A\nput replaces every field, so an absent cover clears the cover. A new event\nis always a draft.",
        "required": [
          "name",
          "startsAt",
          "endsAt"
        ],
        "properties": {
          "description": {
            "$ref": "#/components/schemas/Description"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "games": {
            "$ref": "#/components/schemas/Games"
          },
          "image": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ImageName"
              }
            ]
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "name": {
            "$ref": "#/components/schemas/EventName"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "EventId": {
        "type": "string",
        "format": "uuid"
      },
      "EventName": {
        "type": "string",
        "maxLength": 60,
        "minLength": 1
      },
      "EventStatus": {
        "type": "string",
        "description": "`Draft` is visible to admins only, and its code opens nothing.",
        "enum": [
          "draft",
          "published"
        ]
      },
      "EventStatusChange": {
        "type": "object",
        "description": "Body of `POST /api/admin/events/{id}/status`.",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          }
        },
        "additionalProperties": false
      },
...
      "Games": {
        "type": "string",
        "maxLength": 200,
        "minLength": 0
      },
...
      "ImageName": {
        "type": "string",
        "description": "A file name under the event images of the site",
        "maxLength": 80,
        "minLength": 1,
        "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$"
      },
...
      "Location": {
        "type": "string",
        "maxLength": 60,
        "minLength": 0
      },
...
      "Paginated_Event": {
        "type": "object",
        "description": "One page of results, and the totals a client needs for a pager.",
        "required": [
          "items",
          "total",
          "page",
          "pageSize",
          "totalPages"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "An event as every client sees it. The check-in code is not here: only an\nadmin reads it, through [`EventDetail`].",
              "required": [
                "id",
                "name",
                "location",
                "games",
                "description",
                "startsAt",
                "endsAt",
                "status",
                "checkinCount",
                "createdAt"
              ],
              "properties": {
                "checkinCount": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "description": {
                  "$ref": "#/components/schemas/Description"
                },
                "endsAt": {
                  "type": "string",
                  "format": "date-time",
                  "description": "The last game."
                },
                "games": {
                  "$ref": "#/components/schemas/Games"
                },
                "id": {
                  "$ref": "#/components/schemas/EventId"
                },
                "image": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/ImageName"
                    }
                  ]
                },
                "location": {
                  "$ref": "#/components/schemas/Location"
                },
                "name": {
                  "$ref": "#/components/schemas/EventName"
                },
                "startsAt": {
                  "type": "string",
                  "format": "date-time",
                  "description": "The doors open."
                },
                "status": {
                  "$ref": "#/components/schemas/EventStatus"
                }
              }
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 1
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Every row that matches, ignoring the window.",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "PlayerId": {
        "type": "string",
        "format": "uuid"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}